refactor(scene3d): 抽出相机控制器到 scene3dViewControllers.tsx [split 4/N]
Scene3DControls/InitialCameraPose/FocusController/CaptureBinder/CameraViewEditController
(含取景记录器 useFrame 脏判断)移出;壳 2230→1805。双 tsc 零错误
